+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,7 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+- ui: fix height-flickering of list entry when new drawing is lazy-loaded into carousel
 
 0.4.2
 =====